streaming/ray/requirements.txt
ray
redis
lru-dict